python-src/pyproject.toml
[project]
name = "llm-style-terrain"
version = "0.1.0"
description = "Perlin noise heightmap generation with SDXL image-to-image refinement"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "noise>=1.0.2",
    "pillow>=9.0.0",
    "torch>=2.0.0",
    "diffusers>=0.20.0",
    "accelerate>=0.20.0",
    "transformers>=4.30.0",
]
